pyproject.toml
[tool.poetry]
name = "fastapi-backend-template"
version = "0.4.1"
description = "A FastAPI backend template."
authors = ["ArielMAJ <ariel.maj@hotmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
loguru = "^0.7.2"
httpx = "^0.26.0"
fastapi-cache2 = "^0.2.1"
alembic = "^1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
fastapi-async-sqlalchemy = "^0.6.1"
psycopg = "^3.2.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.6.0"
pre-commit = "^3.6.0"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
coverage = "^7.4.1"
pytest-asyncio = "^0.23.5"
pytest-mock = "^3.12.0"

[tool.coverage.run]
branch=true
source = ["api"]
omit = [
    "api/__main__.py",
    "api/config.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
